Complete Version 2

+++ Backend/CoffeeBook/Controllers/NewsController.cs
@@ -8,6 +8,7 @@ using Microsoft.Extensions.Configuration;
 using System;
 using System.Collections.Generic;
 using System.Data;
+using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -15,7 +16,6 @@ namespace CoffeeBook.Controllers
 {
     /*[Route("api/[controller]")]*/
     [ApiController]
-    /*[Authorize]*/
     public class NewsController : ControllerBase
     {
         private readonly IConfiguration _config;
@@ -49,32 +49,69 @@ namespace CoffeeBook.Controllers
             return new JsonResult(news);
         }
 
+        [Authorize(Roles = "1")]
         [Route("news/add")]
         [HttpPost]
         public ActionResult Post(News news)
         {
-            int res = service.save(news);
-            if (res > 0) return Ok();
+            string jwt = Request.Cookies["jwt"];
+            if (!string.IsNullOrEmpty(jwt))
+            {
+                var Role = getCurrentRole(jwt);
+                if (Role == "1" || Role == "2")
+                {
+                    int res = service.save(news);
+                    if (res > 0) return Ok();
 
-            return BadRequest();
+                    return BadRequest();
+                }
+            }
+            return Unauthorized(new { message = "Bạn không có quyền truy cập" });
         }
 
         [Route("news/edit/{id}")]
         [HttpPut]
         public ActionResult Put(int id, News news)
         {
-            int res = service.update(id, news);
-            if (res > 0) return Ok();
-            return BadRequest();
+            string jwt = Request.Cookies["jwt"];
+            if (!string.IsNullOrEmpty(jwt))
+            {
+                var Role = getCurrentRole(jwt);
+                if (Role == "1" || Role == "2")
+                {
+                    int res = service.update(id, news);
+                    if (res > 0) return Ok();
+                    return BadRequest();
+                }
+            }
+            return Unauthorized(new { message = "Bạn không có quyền truy cập" });
         }
 
         [Route("news/delete/{id}")]
         [HttpDelete]
         public ActionResult Delete(int id)
         {
-            int res = service.deleteById(id);
-            if (res > 0) return Ok();
-            return BadRequest();
+            string jwt = Request.Cookies["jwt"];
+            if (!string.IsNullOrEmpty(jwt))
+            {
+                var Role = getCurrentRole(jwt);
+                if (Role == "1" || Role == "2")
+                {
+                    int res = service.deleteById(id);
+                    if (res > 0) return Ok();
+                    return BadRequest();
+                }
+            }
+            return Unauthorized(new { message = "Bạn không có quyền truy cập" });
+        }
+
+        private string getCurrentRole(string jwt)
+        {
+            var handler = new JwtSecurityTokenHandler();
+            var jsonToken = handler.ReadToken(jwt);
+            var tokenS = jsonToken as JwtSecurityToken;
+            var role = tokenS.Claims.First(claim => claim.Type == "RoleId").Value;
+            return role;
         }
     }
 }
